KAFKA-2509: Replace LeaderAndIsr{Request,Response} with o.a.k.c requests equivalent
